+++ Packages/OsaurusCore/Resources/Localizable.xcstrings
@@ -464,9 +464,6 @@
         }
       }
     },
-    "%lld database(s) didn't migrate" : {
-
-    },
     "%lld files" : {
 
     },
@@ -552,13 +549,6 @@
       }
     },
     "%lld of 24 words" : {
-
-    },
-    "%lld plugin database(s) can't be opened with the current encryption key. They're almost always left over from uninstalled plugins or development test runs. Removing them deletes the corresponding folders under ~/.osaurus/Tools/. Real plugin data won't be touched." : {
-      "comment" : "An alert message explaining that a user can remove plugin data that's left over from uninstalled plugins or development test runs. The number in the plural form is replaced with the actual number of such plugin DBs.",
-      "isCommentAutoGenerated" : true
-    },
-    "%lld plugin database(s) can't be opened with the current key. These are typically left behind when you uninstall a plugin. Your chats and configuration are unaffected." : {
 
     },
     "%lld plugin%@ need%@ system permissions" : {
@@ -2141,9 +2131,6 @@
       "isCommentAutoGenerated" : true
     },
     "Always listening for wake words" : {
-
-    },
-    "An earlier build encrypted these by mistake. They're now back as plaintext where the app expects them." : {
 
     },
     "and %lld more..." : {
@@ -4116,9 +4103,6 @@
     "CLAUDE.md" : {
 
     },
-    "Clean up orphaned plugin data" : {
-
-    },
     "Clear" : {
       "localizations" : {
         "de" : {
@@ -5444,9 +5428,6 @@
           }
         }
       }
-    },
-    "Core data can't be decrypted" : {
-
     },
     "Core Model" : {
       "localizations" : {
@@ -15008,9 +14989,6 @@
     "Or type a custom answer…" : {
 
     },
-    "Orphaned plugin data" : {
-
-    },
     "Osaurus" : {
       "shouldTranslate" : false
     },
@@ -15924,9 +15902,6 @@
           }
         }
       }
-    },
-    "Plaintext copies are preserved under ~/.osaurus/.pre-encryption-backup/. Re-launch Osaurus to retry, or use Export plaintext backup below to bundle them." : {
-
     },
     "Play" : {
 
@@ -18253,9 +18228,6 @@
     "Remove old" : {
 
     },
-    "Remove orphaned plugin data?" : {
-
-    },
     "Remove Plugin?" : {
       "localizations" : {
         "de" : {
@@ -18847,9 +18819,6 @@
     "Restore Anyway" : {
 
     },
-    "Restore the original Keychain entry to recover. Rotating the key in this state would permanently destroy the unreadable data." : {
-
-    },
     "Restore View Defaults" : {
       "localizations" : {
         "de" : {
@@ -18881,9 +18850,6 @@
           }
         }
       }
-    },
-    "Restored %lld configuration file(s)" : {
-
     },
     "Restored the database" : {
 
@@ -20041,9 +20007,6 @@
     "Secrets are injected as environment variables when this agent runs commands or plugins in the sandbox." : {
 
     },
-    "Securing your data" : {
-
-    },
     "Security update is ready" : {
 
     },
@@ -20864,9 +20827,6 @@
     "Show details" : {
       "comment" : "Text that appears in a button to show advanced details about a feature.",
       "isCommentAutoGenerated" : true
-    },
-    "Show details (%lld)" : {
-
     },
     "Show in chat" : {
 
@@ -25459,9 +25419,6 @@
 
     },
     "Your data stays yours" : {
-
-    },
-    "Your existing data was encrypted with a different key — typically because the macOS Keychain was wiped, restored from a different machine, or a previous app version stored a different key. Osaurus is opening it in degraded mode." : {
 
     },
     "Your identity is active" : {
