+++ Packages/OsaurusCore/Storage/MethodDatabase.swift
@@ -65,7 +65,7 @@ public final class MethodDatabase: @unchecked Sendable {
 
     public func open() throws {
         // See `ChatHistoryDatabase.open()` for the gate rationale.
-        StorageMigrationCoordinator.blockingAwaitReady()
+        StorageMutationGate.blockingAwaitNotMutating()
         try queue.sync {
             guard db == nil else { return }
             OsaurusPaths.ensureExistsSilent(OsaurusPaths.methods())
