+++ Packages/OsaurusCore/Views/Settings/StorageSettingsView.swift
@@ -3,12 +3,9 @@
 //  osaurus
 //
 //  Settings panel for at-rest encryption: explains the encryption
-//  posture in plain language, surfaces real recovery situations
-//  (key mismatch on a core DB) without false alarms (orphaned
-//  plugin DBs from uninstalled plugins / leaked dev tests), and
-//  exposes the two admin actions — export plaintext backup and
-//  rotate the storage key — with guardrails so a user can't
-//  accidentally destroy their data.
+//  posture in plain language and exposes the two admin actions —
+//  export plaintext backup and rotate the storage key — with
+//  guardrails so a user can't accidentally destroy their data.
 //
 //  Surfaced by the WhatsNew page action `openStorageSettings` and
 //  reachable from the management settings sidebar.
@@ -27,12 +24,8 @@ public struct StorageSettingsView: View {
     @State private var showRotateConfirm: Bool = false
     @State private var errorMessage: String?
 
-    @State private var lastOutcome: StorageMigrator.OutcomeSummary?
-    @State private var keyMismatchTargets: [StorageMigrator.DatabaseTarget] = []
     @State private var hasExportedBackupThisSession: Bool = false
-    @State private var showCleanupConfirm: Bool = false
     @State private var showTechnicalDetails: Bool = false
-    @State private var showMismatchDetails: Bool = false
 
     @State private var hasAppeared = false
 
@@ -49,20 +42,6 @@ public struct StorageSettingsView: View {
                 VStack(alignment: .leading, spacing: 24) {
                     aboutCard
                     statusCard
-
-                    if !coreMismatchTargets.isEmpty {
-                        coreKeyMismatchCard
-                    } else if !pluginMismatchTargets.isEmpty {
-                        pluginOrphanCard
-                    }
-
-                    if let outcome = lastOutcome, !outcome.failedTargets.isEmpty {
-                        partialFailureCard(outcome: outcome)
-                    }
-                    if let outcome = lastOutcome, outcome.jsonFilesRecovered > 0 {
-                        recoveryCard(outcome: outcome)
-                    }
-
                     actionsCard
                     footnote
                 }
@@ -90,35 +69,9 @@ public struct StorageSettingsView: View {
         } message: {
             Text(rotateAlertMessage)
         }
-        .alert(
-            "Remove orphaned plugin data?",
-            isPresented: $showCleanupConfirm
-        ) {
-            Button(localized: "Cancel", role: .cancel) {}
-            Button(localized: "Remove", role: .destructive) { cleanupOrphans() }
-        } message: {
-            Text(
-                "\(pluginMismatchTargets.count) plugin database(s) can't be opened with the current encryption key. They're almost always left over from uninstalled plugins or development test runs. Removing them deletes the corresponding folders under ~/.osaurus/Tools/. Real plugin data won't be touched."
-            )
-        }
     }
 
     // MARK: - Derived state
-
-    private var coreMismatchTargets: [StorageMigrator.DatabaseTarget] {
-        keyMismatchTargets.filter { $0.pluginId == nil }
-    }
-
-    private var pluginMismatchTargets: [StorageMigrator.DatabaseTarget] {
-        keyMismatchTargets.filter { $0.pluginId != nil }
-    }
-
-    /// Rotation is unsafe when a core DB is in mismatch — the rotate
-    /// path would re-encrypt the unreadable file under the new key
-    /// and delete the chance to ever recover it. The button stays
-    /// visible so the user knows the action exists, but the alert
-    /// path is short-circuited.
-    private var rotateBlocked: Bool { !coreMismatchTargets.isEmpty }
 
     private var rotateAlertMessage: String {
         if hasExportedBackupThisSession {
@@ -260,237 +213,7 @@ public struct StorageSettingsView: View {
         if !keyPresent {
             return "Generate a key from the Keychain to encrypt new data."
         }
-        if !coreMismatchTargets.isEmpty {
-            return "Core data needs your attention — see the recovery card below."
-        }
-        if !pluginMismatchTargets.isEmpty {
-            return "Your data is encrypted at rest. Stale plugin data is listed below."
-        }
         return "Your data is encrypted at rest. No action needed."
-    }
-
-    // MARK: - Core key-mismatch card (loud red)
-
-    /// Loud recovery card shown when one of the four core databases
-    /// (chat history / memory / methods / tool index) can't be
-    /// opened with the current Keychain key. This is the genuine
-    /// "your data is encrypted with a different key than what's in
-    /// my Keychain right now" situation. We block destructive
-    /// actions (rotate) until the user takes a corrective step.
-    private var coreKeyMismatchCard: some View {
-        VStack(alignment: .leading, spacing: 12) {
-            Label {
-                Text("Core data can't be decrypted", bundle: .module)
-            } icon: {
-                Image(systemName: "exclamationmark.triangle.fill")
-                    .foregroundColor(theme.errorColor)
-            }
-            .font(.system(size: 13, weight: .semibold))
-            .foregroundColor(theme.primaryText)
-
-            Text(
-                "Your existing data was encrypted with a different key — typically because the macOS Keychain was wiped, restored from a different machine, or a previous app version stored a different key. Osaurus is opening it in degraded mode.",
-                bundle: .module
-            )
-            .font(.system(size: 12))
-            .foregroundColor(theme.secondaryText)
-            .fixedSize(horizontal: false, vertical: true)
-
-            VStack(alignment: .leading, spacing: 6) {
-                ForEach(coreMismatchTargets, id: \.label) { target in
-                    HStack(spacing: 8) {
-                        Image(systemName: "lock.trianglebadge.exclamationmark")
-                            .font(.system(size: 11))
-                            .foregroundColor(theme.errorColor)
-                        Text(target.label.capitalized)
-                            .font(.system(size: 12, weight: .medium))
-                            .foregroundColor(theme.primaryText)
-                    }
-                }
-            }
-
-            Text(
-                "Restore the original Keychain entry to recover. Rotating the key in this state would permanently destroy the unreadable data.",
-                bundle: .module
-            )
-            .font(.system(size: 11))
-            .foregroundColor(theme.secondaryText)
-            .fixedSize(horizontal: false, vertical: true)
-        }
-        .padding(16)
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .background(
-            RoundedRectangle(cornerRadius: 12)
-                .fill(theme.errorColor.opacity(0.08))
-                .overlay(
-                    RoundedRectangle(cornerRadius: 12)
-                        .stroke(theme.errorColor.opacity(0.4), lineWidth: 1)
-                )
-        )
-    }
-
-    // MARK: - Plugin orphan card (quiet warning)
-
-    /// Quieter informational card shown when only **plugin** DBs
-    /// fail to decrypt. These are almost always left behind by
-    /// uninstalled plugins or leaked development test runs (the
-    /// `com.test.*` filter in `StorageMigrator.databaseTargets()`
-    /// catches the historical dev-leak pattern; this card is the
-    /// safety net for everything else). Loss of this data is
-    /// tolerable, so we offer a one-click cleanup instead of a
-    /// scary alarm.
-    private var pluginOrphanCard: some View {
-        VStack(alignment: .leading, spacing: 12) {
-            Label {
-                Text("Orphaned plugin data", bundle: .module)
-            } icon: {
-                Image(systemName: "tray.full")
-                    .foregroundColor(theme.warningColor)
-            }
-            .font(.system(size: 13, weight: .semibold))
-            .foregroundColor(theme.primaryText)
-
-            Text(
-                "\(pluginMismatchTargets.count) plugin database(s) can't be opened with the current key. These are typically left behind when you uninstall a plugin. Your chats and configuration are unaffected.",
-                bundle: .module
-            )
-            .font(.system(size: 12))
-            .foregroundColor(theme.secondaryText)
-            .fixedSize(horizontal: false, vertical: true)
-
-            DisclosureGroup(isExpanded: $showMismatchDetails) {
-                VStack(alignment: .leading, spacing: 4) {
-                    ForEach(pluginMismatchTargets, id: \.label) { target in
-                        Text(target.label)
-                            .font(.system(size: 11, design: .monospaced))
-                            .foregroundColor(theme.tertiaryText)
-                            .textSelection(.enabled)
-                    }
-                }
-                .padding(.top, 6)
-            } label: {
-                Text("Show details (\(pluginMismatchTargets.count))", bundle: .module)
-                    .font(.system(size: 11, weight: .medium))
-                    .foregroundColor(theme.tertiaryText)
-            }
-            .accentColor(theme.tertiaryText)
-
-            Button {
-                showCleanupConfirm = true
-            } label: {
-                HStack(spacing: 6) {
-                    Image(systemName: "trash")
-                        .font(.system(size: 11, weight: .semibold))
-                    Text("Clean up orphaned plugin data", bundle: .module)
-                        .font(.system(size: 12, weight: .semibold))
-                }
-                .foregroundColor(.white)
-                .padding(.horizontal, 12)
-                .padding(.vertical, 7)
-                .background(
-                    RoundedRectangle(cornerRadius: 6)
-                        .fill(theme.warningColor)
-                )
-            }
-            .buttonStyle(PlainButtonStyle())
-            .disabled(isWorking)
-        }
-        .padding(16)
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .background(
-            RoundedRectangle(cornerRadius: 12)
-                .fill(theme.warningColor.opacity(0.06))
-                .overlay(
-                    RoundedRectangle(cornerRadius: 12)
-                        .stroke(theme.warningColor.opacity(0.25), lineWidth: 1)
-                )
-        )
-    }
-
-    // MARK: - Migration outcome cards
-
-    /// Surfaces databases the migrator couldn't re-encrypt. The
-    /// originals are still on disk under
-    /// `~/.osaurus/.pre-encryption-backup/` so the user can recover
-    /// or retry. Without this card the user has no way of knowing
-    /// they're running in a partially-degraded mode.
-    private func partialFailureCard(outcome: StorageMigrator.OutcomeSummary) -> some View {
-        VStack(alignment: .leading, spacing: 12) {
-            Label {
-                Text("\(outcome.failedTargets.count) database(s) didn't migrate", bundle: .module)
-            } icon: {
-                Image(systemName: "exclamationmark.triangle.fill")
-                    .foregroundColor(theme.warningColor)
-            }
-            .font(.system(size: 13, weight: .semibold))
-            .foregroundColor(theme.primaryText)
-
-            ForEach(outcome.failedTargets.sorted(by: { $0.key < $1.key }), id: \.key) { item in
-                VStack(alignment: .leading, spacing: 2) {
-                    Text(item.key)
-                        .font(.system(size: 12, weight: .medium))
-                        .foregroundColor(theme.primaryText)
-                    Text(item.value)
-                        .font(.system(size: 11))
-                        .foregroundColor(theme.secondaryText)
-                }
-            }
-
-            Text(
-                "Plaintext copies are preserved under ~/.osaurus/.pre-encryption-backup/. Re-launch Osaurus to retry, or use Export plaintext backup below to bundle them.",
-                bundle: .module
-            )
-            .font(.system(size: 11))
-            .foregroundColor(theme.secondaryText)
-            .fixedSize(horizontal: false, vertical: true)
-        }
-        .padding(16)
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .background(
-            RoundedRectangle(cornerRadius: 12)
-                .fill(theme.warningColor.opacity(0.08))
-                .overlay(
-                    RoundedRectangle(cornerRadius: 12)
-                        .stroke(theme.warningColor.opacity(0.4), lineWidth: 1)
-                )
-        )
-    }
-
-    /// Shown after a v1→v2 launch where the migrator restored
-    /// `.osec` JSON files back to plaintext (the recovery path —
-    /// see `StorageMigrator.recoverEncryptedJSON`). Quiet
-    /// confirmation that the user's agents/themes/config are back.
-    private func recoveryCard(outcome: StorageMigrator.OutcomeSummary) -> some View {
-        HStack(alignment: .top, spacing: 12) {
-            Image(systemName: "arrow.clockwise.circle.fill")
-                .foregroundColor(theme.accentColor)
-                .font(.system(size: 16))
-            VStack(alignment: .leading, spacing: 2) {
-                Text(
-                    "Restored \(outcome.jsonFilesRecovered) configuration file(s)",
-                    bundle: .module
-                )
-                .font(.system(size: 12, weight: .semibold))
-                .foregroundColor(theme.primaryText)
-                Text(
-                    "An earlier build encrypted these by mistake. They're now back as plaintext where the app expects them.",
-                    bundle: .module
-                )
-                .font(.system(size: 11))
-                .foregroundColor(theme.secondaryText)
-                .fixedSize(horizontal: false, vertical: true)
-            }
-        }
-        .padding(14)
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .background(
-            RoundedRectangle(cornerRadius: 12)
-                .fill(theme.accentColor.opacity(0.06))
-                .overlay(
-                    RoundedRectangle(cornerRadius: 12)
-                        .stroke(theme.accentColor.opacity(0.25), lineWidth: 1)
-                )
-        )
     }
 
     // MARK: - Actions card
@@ -513,8 +236,7 @@ public struct StorageSettingsView: View {
                     subtitle:
                         "Decrypts every artifact under ~/.osaurus and writes a plaintext copy to the destination of your choice. Recommended before reinstalling macOS or moving Macs.",
                     isPrimary: true,
-                    isDisabled: isWorking,
-                    disabledHelp: nil
+                    isDisabled: isWorking
                 ) {
                     runExport(reason: .userInitiated)
                 }
@@ -525,14 +247,9 @@ public struct StorageSettingsView: View {
                     icon: "key.fill",
                     title: "Rotate storage key",
                     buttonLabel: "Rotate",
-                    subtitle: rotateBlocked
-                        ? "Disabled while core data can't be decrypted. Rotating now would permanently destroy the unreadable data."
-                        : "Generate a new 256-bit key and re-encrypt every artifact. The old key is destroyed.",
+                    subtitle: "Generate a new 256-bit key and re-encrypt every artifact. The old key is destroyed.",
                     isPrimary: false,
-                    isDisabled: isWorking || rotateBlocked,
-                    disabledHelp: rotateBlocked
-                        ? "Disabled while a core database can't be decrypted."
-                        : nil
+                    isDisabled: isWorking
                 ) {
                     showRotateConfirm = true
                 }
@@ -565,7 +282,6 @@ public struct StorageSettingsView: View {
         subtitle: String,
         isPrimary: Bool,
         isDisabled: Bool,
-        disabledHelp: String?,
         action: @escaping () -> Void
     ) -> some View {
         HStack(alignment: .top, spacing: 12) {
@@ -584,7 +300,6 @@ public struct StorageSettingsView: View {
                 label: buttonLabel,
                 isPrimary: isPrimary,
                 isDisabled: isDisabled,
-                disabledHelp: disabledHelp,
                 action: action
             )
         }
@@ -596,10 +311,9 @@ public struct StorageSettingsView: View {
         label: String,
         isPrimary: Bool,
         isDisabled: Bool,
-        disabledHelp: String?,
         action: @escaping () -> Void
     ) -> some View {
-        let button = Button(action: action) {
+        Button(action: action) {
             HStack(spacing: 6) {
                 Image(systemName: icon)
                     .font(.system(size: 11, weight: .semibold))
@@ -613,12 +327,6 @@ public struct StorageSettingsView: View {
         }
         .buttonStyle(PlainButtonStyle())
         .disabled(isDisabled)
-
-        if let disabledHelp, isDisabled {
-            button.help(Text(LocalizedStringKey(disabledHelp), bundle: .module))
-        } else {
-            button
-        }
     }
 
     @ViewBuilder
@@ -665,8 +373,6 @@ public struct StorageSettingsView: View {
 
     private func refresh() async {
         keyPresent = StorageKeyManager.shared.keyExists()
-        lastOutcome = await StorageMigrator.shared.loadLastOutcome()
-        keyMismatchTargets = await StorageMigrator.shared.detectKeyMismatch()
     }
 
     /// Why an export is being run — drives the open-panel copy,
@@ -747,20 +453,4 @@ public struct StorageSettingsView: View {
         }
     }
 
-    private func cleanupOrphans() {
-        let toRemove = pluginMismatchTargets
-        isWorking = true
-        errorMessage = nil
-        Task {
-            let summary = await StorageExportService.shared.cleanupOrphanedPluginDatabases(
-                targets: toRemove
-            )
-            await MainActor.run {
-                self.isWorking = false
-                self.lastSummary =
-                    "Removed \(summary.directoriesRemoved) orphaned plugin director\(summary.directoriesRemoved == 1 ? "y" : "ies")."
-            }
-            await refresh()
-        }
-    }
 }
